Add highlighting group helpers and models in one group
