+++ Sources/Models and Helpers/Helpers.swift
@@ -150,28 +150,3 @@ internal struct WeakSet<T: AnyObject>: Sequence {
         }
     }
 }
-
-// TODO: move to own file when it's set.
-@objc public class IndexId: NSObject  {
-    var name: String
-    var id: String
-    
-    public convenience init(name: String) {
-        self.init(name: name, id: "")
-    }
-    
-    public init(name: String, id: String) {
-        self.name = name
-        self.id = id
-    }
-    
-    override public var hashValue: Int {
-        return "\(name)\(id)".hashValue
-    }
-    
-    override public func isEqual(_ object: Any?) -> Bool {
-        guard let rhs = object as? IndexId else { return false }
-        
-        return (name == rhs.name) && (id == rhs.id)
-    }
-}
